refactor: 당첨 번호 유효성 검증 로직 분리
